+++ example/pages/index.tsx
@@ -10,12 +10,15 @@ export const getStaticProps: GetStaticProps = async ({ locale }) => ({
 
 export default function Home() {
   const { t } = useTranslation()
+  const name = 'Sam'
 
   return (
     <>
       <ul>
         <li>1: {t('Hello world')}</li>
         <li>2: {t(`Hello world`)}</li>
+        <li>3: {t(`Happy birthday, {{name}}!`, { name })}</li>
+        <li>4: {t(`Happy birthday, {{ name }}!`, { name })}</li>
       </ul>
 
       <ul>
@@ -78,6 +81,13 @@ export default function Home() {
             vulputate venenatis, ipsum mi efficitur est, vel ornare ex enim eget orci. Nam mattis
             libero quis finibus hendrerit.
           </Trans>
+        </li>
+        {/* prettier-ignore */}
+        <li>
+          13: <Trans t={t}>Happy birthday, {{name}}!</Trans>
+        </li>
+        <li>
+          14: <Trans t={t}>Happy birthday, {{ name }}!</Trans>
         </li>
       </ul>
     </>