+++ tests/astToKey.spec.ts
@@ -21,7 +21,7 @@ function astToKeyFromCode(code: string) {
 
         JSXElement(path: NodePath<BabelTypes.JSXElement>) {
           if (!t.isJSXIdentifier(path.node.openingElement.name, { name: 'Trans' })) return
-          key = astToKey(path.node.children, { code })
+          key = astToKey(path.node.children, { code, jsx: true })
         },
       },
     }
@@ -51,6 +51,23 @@ describe('astToKey (functional)', () => {
     `)
 
     expect(key).toEqual('Sign in to your account')
+  })
+
+  it('handles interpolated variable', () => {
+    const key = astToKeyFromCode(`
+      const name = 'Sam'
+      t(\`Happy birthday, {{name}}!\`, { name })
+    `)
+
+    expect(key).toEqual('Happy birthday, {{name}}!')
+  })
+
+  it('handles multiple interpolated variables', () => {
+    const key = astToKeyFromCode(`
+      t(\`Big thanks to {{a}} and {{ b }}.\`)
+    `)
+
+    expect(key).toEqual('Big thanks to {{a}} and {{ b }}.')
   })
 })
 
@@ -130,6 +147,30 @@ describe('astToKey (JSX)', () => {
     expect(key).toEqual("<0></0> There's something in the water.")
   })
 
+  it('handles interpolated variable', () => {
+    const key = astToKeyFromCode(`
+      <Trans t={t}>Happy birthday, {{name}}!</Trans>
+    `)
+
+    expect(key).toEqual('Happy birthday, {{name}}!')
+  })
+
+  it('handles multiple interpolated variables', () => {
+    const key = astToKeyFromCode(`
+      <Trans t={t}>Big thanks to {{a}} and {{ b }}.</Trans>
+    `)
+
+    expect(key).toEqual('Big thanks to {{a}} and {{ b }}.')
+  })
+
+  it('handles interpolated variable inside of an interpolated component', () => {
+    const key = astToKeyFromCode(`
+      <Trans t={t}>Happy birthday, <Bold>birthday person {{name}}</Bold>!</Trans>
+    `)
+
+    expect(key).toEqual('Happy birthday, <1>birthday person {{name}}</1>!')
+  })
+
   it('errors on interpolated expression', () => {
     expect(() =>
       astToKeyFromCode(`
@@ -157,6 +198,6 @@ describe('astToKey (JSX)', () => {
   it('errors on unknown AST type', () => {
     // eslint-disable-next-line @typescript-eslint/ban-ts-comment
     // @ts-ignore Forcing an invalid AST type
-    expect(() => astToKey([{ type: 'Foobar' }])).toThrowErrorMatchingSnapshot()
+    expect(() => astToKey([{ type: 'Foobar' }], {})).toThrowErrorMatchingSnapshot()
   })
 })