+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 1.3.0
+
+**Features**
+
+- Support variable interpolation for `t` (`t('Foo {{bar}}', { bar })`)
+- Support variable interpolation for `Trans` (`<Trans t={t}>Foo {{bar}}</Trans>`)
+
 # 1.2.0
 
 **Features**